+++ app/ui/settings/panes/pref_obj.h
@@ -22,12 +22,12 @@ typedef struct pref_dropdown_string_entry_t {
 
 lv_obj_t *pref_checkbox(lv_obj_t *parent, const char *title, bool *value, bool reverse);
 
-lv_obj_t *pref_dropdown_int(lv_obj_t *parent, const pref_dropdown_int_entry_t *entries, int num_entries, int *value);
+lv_obj_t *pref_dropdown_int(lv_obj_t *parent, const pref_dropdown_int_entry_t *entries, size_t num_entries, int *value);
 
-lv_obj_t *pref_dropdown_int_pair(lv_obj_t *parent, const pref_dropdown_int_pair_entry_t *entries, int num_entries,
+lv_obj_t *pref_dropdown_int_pair(lv_obj_t *parent, const pref_dropdown_int_pair_entry_t *entries, size_t num_entries,
                                  int *value_a, int *value_b);
 
-lv_obj_t *pref_dropdown_string(lv_obj_t *parent, const pref_dropdown_string_entry_t *entries, int num_entries,
+lv_obj_t *pref_dropdown_string(lv_obj_t *parent, const pref_dropdown_string_entry_t *entries, size_t num_entries,
                                char **value);
 
 lv_obj_t *pref_slider(lv_obj_t *parent, int *value, int min, int max, int step);
